# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RealEstate AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai/conversations:
    get:
      operationId: listConversations
      tags: [aiChat]
      summary: List user's AI conversations
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiConversation"
    post:
      operationId: createConversation
      tags: [aiChat]
      summary: Start a new AI conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiConversation"
  /ai/conversations/{id}:
    get:
      operationId: getConversation
      tags: [aiChat]
      summary: Get a conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiConversationWithMessages"
    delete:
      operationId: deleteConversation
      tags: [aiChat]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /ai/conversations/{id}/messages:
    post:
      operationId: sendMessage
      tags: [aiChat]
      summary: Send a message to the AI assistant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiMessageInput"
      responses:
        "200":
          description: AI response with structured data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiMessage"

  # ── Investment Interests ───────────────────────────────────────────────────
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
  /admin/users/{id}/role:
    patch:
      operationId: updateUserRole
      tags: [admin]
      summary: Update a user's role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /admin/investment-interests/{id}/status:
    patch:
      operationId: updateInvestmentInterestStatus
      tags: [admin]
      summary: Update investment interest status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterestStatusUpdate"
      responses:
        "200":
          description: Updated interest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvestmentInterest"

  # ── Object Storage ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        fullName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin]
        investorType:
          type: ["string", "null"]
          enum: [fix_flip, rental, development, mixed, null]
        bio:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, clerkId, email, role, createdAt]
...
    RoleUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [user, admin]
      required: [role]

    # ── Properties ───────────────────────────────────────────────────────────
...
    ProjectPhase:
      type: object
      properties:
        name:
          type: string
        progress:
          type: integer
        status:
          type: string
          enum: [pending, active, completed]
      required: [name, progress, status]
    Project:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        location:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
        status:
          type: string
          enum: [preventa, construccion, entrega, completado]
        roiEstimated:
          type: ["number", "null"]
        totalInvestment:
          type: ["number", "null"]
        raisedAmount:
          type: ["number", "null"]
        phases:
          type: array
          items:
            $ref: "#/components/schemas/ProjectPhase"
        isPublic:
          type: boolean
        images:
          type: array
          items:
            type: string
        investorCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, userId, title, location, status, phases, isPublic, images, investorCount, createdAt]
...
    AiConversation:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, userId, title, createdAt]
    AiConversationInput:
      type: object
      properties:
        title:
          type: string
      required: [title]
    AiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        structuredData:
          type: ["object", "null"]
          description: Structured ARV/analysis data for rich card rendering
        createdAt:
          type: string
          format: date-time
      required: [id, conversationId, role, content, createdAt]
    AiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/AiMessage"
      required: [id, userId, title, createdAt, messages]
    AiMessageInput:
      type: object
      properties:
        content:
          type: string
      required: [content]

    # ── Investment Interests ───────────────────────────────────────────────────
    InvestmentInterest:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        projectId:
          type: integer
        amountInterested:
          type: ["number", "null"]
        message:
          type: ["string", "null"]
        status:
          type: string
          enum: [pendiente, contactado, rechazado]
        createdAt:
          type: string
          format: date-time
        project:
          $ref: "#/components/schemas/Project"
        user:
          $ref: "#/components/schemas/User"
      required: [id, userId, projectId, status, createdAt, project, user]
...
    InterestStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pendiente, contactado, rechazado]
      required: [status]

    # ── Notifications ─────────────────────────────────────────────────────────
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [inspection_due, investment_update, project_update, system]
        title:
          type: string
        body:
          type: string
        read:
          type: boolean
        entityId:
          type: ["integer", "null"]
        entityType:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, userId, type, title, body, read, createdAt]
...