# lib/api-zod/src/generated/types/aiConversationWithMessages.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RealEstate AI API
 * OpenAPI spec version: 0.1.0
 */
import type { AiMessage } from './aiMessage';

export interface AiConversationWithMessages {
  id: number;
  userId: number;
  title: string;
  createdAt: Date;
  messages: AiMessage[];
}