# lib/api-zod/src/generated/types/projectPhase.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RealEstate AI API
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectPhaseStatus } from './projectPhaseStatus';

export interface ProjectPhase {
  name: string;
  progress: number;
  status: ProjectPhaseStatus;
}