# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RealEstate AI API
 * OpenAPI spec version: 0.1.0
 */
import type { UserInvestorType } from './userInvestorType';
import type { UserRole } from './userRole';

export interface User {
  id: number;
  clerkId: string;
  email: string;
  /** @nullable */
  fullName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  role: UserRole;
  /** @nullable */
  investorType?: UserInvestorType;
  /** @nullable */
  bio?: string | null;
  createdAt: Date;
}