# lib/db/src/schema/users.ts
import { pgTable, text, serial, timestamp, jsonb } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const usersTable = pgTable("users", {
  id: serial("id").primaryKey(),
  clerkId: text("clerk_id").notNull().unique(),
  email: text("email").notNull().unique(),
  fullName: text("full_name"),
  avatarUrl: text("avatar_url"),
  role: text("role", { enum: ["user", "admin"] }).notNull().default("user"),
  investorType: text("investor_type", { enum: ["fix_flip", "rental", "development", "mixed"] }),
  bio: text("bio"),
  notificationPrefs: jsonb("notification_prefs").$type<{
    nuevasZonas: boolean;
    actualizacionesProyecto: boolean;
    alertasPrecios: boolean;
    resumenSemanal: boolean;
  }>(),
  onboardingCompleted: text("onboarding_completed"),
  createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
  updatedAt: timestamp("updated_at", { withTimezone: true }).notNull().defaultNow().$onUpdate(() => new Date()),
});

export const insertUserSchema = createInsertSchema(usersTable).omit({ id: true, createdAt: true, updatedAt: true });
export type InsertUser = z.infer<typeof insertUserSchema>;
export type User = typeof usersTable.$inferSelect;